# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Household Finance API
  version: 0.1.0
  description: |
    HTTP API for the self-hosted household finance app. All authenticated routes expect
    `Authorization: Bearer <JWT>` unless noted. See `docs/API_INDEX.md` for prose topic guides.
servers:
  - url: /
    description: Relative to configured API base URL
...
paths:
...
  /categories/rules:
    get:
      tags: [categories]
      summary: List category rules
      security: [{ bearerAuth: [] }]
    post:
      tags: [categories]
      summary: Create category rule
      description: >-
        Body may include amountScope (any, credit_only, debit_only); defaults to any.
        Matches multi-pattern field `patterns` for batch of rules sharing the same scope.
      security: [{ bearerAuth: [] }]
  /categories/rules/household:
    delete:
      tags: [categories]
      summary: Delete all household category rules
      description: >-
        Removes every row in category_rule for this household. Does not affect built-in global rules.
      security: [{ bearerAuth: [] }]
  /categories/rules/bulk:
    post:
      tags: [categories]
      summary: Bulk create household category rules
      description: >-
        Best-effort batch create. Each rule needs pattern, matchType, and categoryId or categoryPath.
        Optional amountScope per row (defaults to any). Response 200 with created rules and per-row errors.
      security: [{ bearerAuth: [] }]
  /categories/rules/builtin/bulk:
    post:
      tags: [categories]
      summary: Bulk create global built-in category rules (owner or admin)
      description: >-
        Same as household bulk but amountScope per row; category must be a default global leaf.
        Response 200 with created rules and per-row errors.
      security: [{ bearerAuth: [] }]
  /categories/rules/{ruleId}:
    patch:
      tags: [categories]
      summary: Update rule
      security: [{ bearerAuth: [] }]
    delete:
      tags: [categories]
      summary: Delete household category rule
      description: Permanently removes a household rule (built-in rules use /categories/rules/builtin/{id}).
      security: [{ bearerAuth: [] }]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT